implemtacion de h2, relacion entre mascotas y dueños

+++ demo/src/main/java/com/doggydr/demo/servicio/ClientService.java
@@ -5,7 +5,7 @@ import java.util.Collection;
 import com.doggydr.demo.entidad.Client;
 
 public interface ClientService {
-    public Client SearchById(int id);
+    public Client SearchById(Long id);
 
     public Collection<Client> SearchAll();
 
@@ -13,7 +13,7 @@ public interface ClientService {
 
     public void Register(Client client);
     
-    public void DeleteById(int id);
+    public void DeleteById(Long id);
 
     public void update(Client client);
 }
